Parking Application with JFrame
A simple parking application written in JFrame. This is a project that I did to try myself when I was learning the Java programming language. The information entered by the user is saved in the TextFile and then the parking fee is calculated. If you wish, you can see the registered vehicles by clicking on the See Garage option from the Main Menu